** feat : Intégration API Auth - Reformatation des codes frontend register et login - Adaptation du Backend AuthService et AuthController pour retourner lors du login et le Token et le User
