+++ backend/Program.cs
@@ -100,6 +100,24 @@ builder.Services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
 
 builder.Services.AddAuthorization();
 
+// Injection des services métiers
+builder.Services.AddScoped<IAuthService, AuthService>();
+builder.Services.AddScoped<IOffreService, OffreService>();
+builder.Services.AddScoped<IEntrepriseService, EntrepriseService>();
+builder.Services.AddScoped<ICandidatureService, CandidatureService>();
+builder.Services.AddScoped<IValidationService, ValidationService>();
+
+// Injection des repositories
+builder.Services.AddScoped<IOffreRepository, OffreRepository>();
+builder.Services.AddScoped<IUserRepository, UserRepository>();
+builder.Services.AddScoped<IEntrepriseRepository, EntrepriseRepository>();
+builder.Services.AddScoped<ICandidatureRepository, CandidatureRepository>();
+builder.Services.AddScoped<IValidationRepository, ValidationRepository>();
+
+// ------------------------------------
+// CONSTRUCTION DE L'APPLICATION
+// ------------------------------------
+
 var app = builder.Build();
 
 // Middleware de développement
@@ -134,11 +152,12 @@ app.MapControllers();
 // Route simple d’accueil
 app.MapGet("/", () => "Bienvenue sur la plateforme de stages !");
 
-// Appliquer les migrations au démarrage
+// Applique les migrations automatiques (optionnel mais pratique en dev)
 using (var scope = app.Services.CreateScope())
 {
-    var dbContext = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
-    dbContext.Database.Migrate();
+    var db = scope.ServiceProvider.GetRequiredService<ApplicationDbContext>();
+    db.Database.Migrate(); // Applique automatiquement les migrations en attente
 }
 
+// Lancement de l'application
 app.Run();
