+++ backend/Services/Interfaces/IAuthService.cs
@@ -18,7 +18,8 @@ namespace backend.Services.Interfaces
         /// </summary>
         /// <param name="dto">The login credentials.</param>
         /// <returns>A JWT token if login is successful, null otherwise.</returns>
-        Task<string?> LoginAsync(LoginDTO dto);
+        //Task<string?> LoginAsync(LoginDTO dto);
+        Task<AuthResultDTO?> LoginAsync(LoginDTO dto);
 
         /// <summary>
         /// Checks if an email is already registered.
