+++ frontend/src/app/register/page.tsx
@@ -4,440 +4,19 @@ import { useRouter } from 'next/navigation';
 import Link from 'next/link';
 import { authService } from '@/services/authService';
 import { Eye, EyeOff, User, Mail, Phone, Lock, CheckCircle, XCircle } from 'lucide-react';
+import AuthRightPanel from '@/components/studentLayouts/authRightPanel';
+import RegisterForm from '@/components/studentLayouts/RegisterForm';
 
-interface RegisterFormData {
-  nom: string;
-  prenom: string;
-  email: string;
-  password: string;
-  confirmPassword: string;
-  telephone: string; 
-}
-
-interface PasswordStrength {
-  score: number;
-  label: string;
-  color: string;
-}
 
 export default function RegisterPage() {
-  const [formData, setFormData] = useState<RegisterFormData>({
-    nom: '',
-    prenom: '',
-    email: '',
-    password: '',
-    confirmPassword: '',
-    telephone: ''  
-  });
-  const [error, setError] = useState('');
-  const [success, setSuccess] = useState('');
-  const [loading, setLoading] = useState(false);
-  const [showPassword, setShowPassword] = useState(false);
-  const [showConfirmPassword, setShowConfirmPassword] = useState(false);
-  const [fieldErrors, setFieldErrors] = useState<Partial<RegisterFormData>>({});
-  const [touchedFields, setTouchedFields] = useState<Set<string>>(new Set());
-  const router = useRouter();
-
-  // Validation des champs en temps réel
-  const validateField = (name: string, value: string): string => {
-    switch (name) {
-      case 'nom':
-      case 'prenom':
-        return value.length < 2 ? 'Minimum 2 caractères requis' : '';
-      case 'email':
-        const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-        return !emailRegex.test(value) ? 'Format d\'email invalide' : '';
-      case 'telephone':
-        const phoneRegex = /^[0-9]{8}$/;
-        return !phoneRegex.test(value) ? 'Format: 8 chiffres (ex: 90000000)' : '';
-      case 'password':
-        if (value.length < 8) return 'Minimum 8 caractères';
-        if (!/(?=.*[a-z])/.test(value)) return 'Au moins une minuscule';
-        if (!/(?=.*[A-Z])/.test(value)) return 'Au moins une majuscule';
-        if (!/(?=.*\d)/.test(value)) return 'Au moins un chiffre';
-        return '';
-      case 'confirmPassword':
-        return value !== formData.password ? 'Les mots de passe ne correspondent pas' : '';
-      default:
-        return '';
-    }
-  };
-
-  // Calcul de la force du mot de passe
-  const getPasswordStrength = (password: string): PasswordStrength => {
-    let score = 0;
-    if (password.length >= 8) score++;
-    if (/(?=.*[a-z])/.test(password)) score++;
-    if (/(?=.*[A-Z])/.test(password)) score++;
-    if (/(?=.*\d)/.test(password)) score++;
-    if (/(?=.*[!@#$%^&*])/.test(password)) score++;
-
-    const strengthMap = {
-      0: { label: 'Très faible', color: 'bg-red-500' },
-      1: { label: 'Faible', color: 'bg-red-400' },
-      2: { label: 'Moyen', color: 'bg-yellow-500' },
-      3: { label: 'Bon', color: 'bg-blue-500' },
-      4: { label: 'Fort', color: 'bg-green-500' },
-      5: { label: 'Très fort', color: 'bg-green-600' }
-    };
-
-    return { score, ...strengthMap[score as keyof typeof strengthMap] };
-  };
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    
-    setFormData(prev => ({
-      ...prev,
-      [name]: value
-    }));
-
-    // Validation en temps réel pour les champs touchés
-    if (touchedFields.has(name)) {
-      const error = validateField(name, value);
-      setFieldErrors(prev => ({
-        ...prev,
-        [name]: error
-      }));
-    }
-
-    setError('');
-    setSuccess('');
-  };
-
-  const handleBlur = (e: React.FocusEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setTouchedFields(prev => new Set(prev).add(name));
-    
-    const error = validateField(name, value);
-    setFieldErrors(prev => ({
-      ...prev,
-      [name]: error
-    }));
-  };
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setError('');
-    setSuccess('');
-
-    // Validation complète
-    const errors: Partial<RegisterFormData> = {};
-    Object.keys(formData).forEach(key => {
-      const error = validateField(key, formData[key as keyof RegisterFormData]);
-      if (error) errors[key as keyof RegisterFormData] = error;
-    });
-
-    if (Object.keys(errors).length > 0) {
-      setFieldErrors(errors);
-      setError('Veuillez corriger les erreurs dans le formulaire');
-      return;
-    }
-
-    setLoading(true);
-
-    try {
-      const response = await authService.register(formData);
-      if (response.success) {
-        setSuccess('Compte créé avec succès ! Redirection en cours...');
-        setTimeout(() => {
-          router.push('/login?registered=true');
-        }, 2000);
-      } else {
-        setError(response.message || 'Erreur lors de l\'inscription');
-      }
-    } catch (err) {
-      setError('Une erreur est survenue. Veuillez réessayer plus tard.');
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const passwordStrength = getPasswordStrength(formData.password);
 
   return (
     <div className="min-h-screen flex flex-col md:flex-row bg-gray-50">
       {/* Panneau latéral */}
-      <div className="hidden md:flex md:w-1/2 bg-gradient-to-br from-blue-600 via-blue-700 to-indigo-800 items-center justify-center p-12 relative overflow-hidden">
-        <div className="absolute inset-0 bg-black opacity-10"></div>
-        <div className="absolute top-0 left-0 w-full h-full">
-          <div className="absolute top-20 left-20 w-32 h-32 bg-white opacity-10 rounded-full animate-pulse"></div>
-          <div className="absolute bottom-32 right-16 w-24 h-24 bg-white opacity-5 rounded-full animate-pulse delay-1000"></div>
-          <div className="absolute top-1/2 right-1/3 w-16 h-16 bg-white opacity-10 rounded-full animate-pulse delay-500"></div>
-        </div>
-        <div className="max-w-md text-white text-center relative z-10">
-          <div className="mb-8">
-            <div className="w-16 h-16 bg-white bg-opacity-20 rounded-full flex items-center justify-center mx-auto mb-4">
-              <svg className="w-8 h-8 text-white" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 13.255A23.931 23.931 0 0112 15c-3.183 0-6.22-.62-9-1.745M16 6V4a2 2 0 00-2-2h-4a2 2 0 00-2-2v2m8 0V6a2 2 0 012 2v6M8 6V4a2 2 0 012-2h4a2 2 0 012 2v2M8 6v12a2 2 0 002 2h4a2 2 0 002-2V6" />
-              </svg>
-            </div>
-          </div>
-          <h1 className="text-4xl font-bold mb-6 leading-tight">Rejoignez-nous</h1>
-          <p className="text-xl opacity-90 mb-8">Créez votre compte étudiant pour accéder aux offres de stage</p>
-          <div className="flex items-center justify-center space-x-6 text-sm opacity-75">
-            <div className="flex items-center">
-              <CheckCircle className="w-4 h-4 mr-2" />
-              <span>Inscription rapide</span>
-            </div>
-            <div className="flex items-center">
-              <CheckCircle className="w-4 h-4 mr-2" />
-              <span>Accès immédiat</span>
-            </div>
-          </div>
-        </div>
-      </div>
+      <AuthRightPanel />
 
       {/* Formulaire */}
-      <div className="w-full md:w-1/2 flex items-center justify-center p-8">
-        <div className="max-w-md w-full">
-          <div className="p-4 sm:p-8">
-            <div className="text-center mb-8">
-              <h2 className="text-3xl font-bold text-gray-900 mb-2">
-                Création de compte
-              </h2>
-              <p className="text-gray-600">Remplissez vos informations personnelles</p>
-            </div>
-
-            {/* Messages d'alerte */}
-            {error && (
-              <div className="mb-6 bg-red-50 border-l-4 border-red-400 p-4 rounded-r-lg">
-                <div className="flex items-center">
-                  <XCircle className="w-5 h-5 text-red-400 mr-2" />
-                  <p className="text-sm text-red-700">{error}</p>
-                </div>
-              </div>
-            )}
-
-            {success && (
-              <div className="mb-6 bg-green-50 border-l-4 border-green-400 p-4 rounded-r-lg">
-                <div className="flex items-center">
-                  <CheckCircle className="w-5 h-5 text-green-400 mr-2" />
-                  <p className="text-sm text-green-700">{success}</p>
-                </div>
-              </div>
-            )}
-
-            <form onSubmit={handleSubmit} className="space-y-6">
-              {/* Nom et Prénom */}
-              <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
-                <div>
-                  <label htmlFor="nom" className="block text-sm font-medium text-gray-700 mb-2">
-                    Nom *
-                  </label>
-                  <div className="relative">
-                    <input
-                      id="nom"
-                      name="nom"
-                      type="text"
-                      required
-                      className={`w-full px-4 py-3 pl-10 border rounded-lg shadow-sm focus:outline-none focus:ring-2 focus:ring-blue-500 transition-colors ${
-                        fieldErrors.nom ? 'border-red-300 focus:border-red-500' : 'border-gray-300 focus:border-blue-500'
-                      }`}
-                      value={formData.nom}
-                      onChange={handleChange}
-                      onBlur={handleBlur}
-                      placeholder="Votre nom"
-                    />
-                    <User className="absolute left-3 top-3.5 h-5 w-5 text-gray-400" />
-                  </div>
-                  {fieldErrors.nom && (
-                    <p className="mt-1 text-sm text-red-600">{fieldErrors.nom}</p>
-                  )}
-                </div>
-
-                <div>
-                  <label htmlFor="prenom" className="block text-sm font-medium text-gray-700 mb-2">
-                    Prénom *
-                  </label>
-                  <div className="relative">
-                    <input
-                      id="prenom"
-                      name="prenom"
-                      type="text"
-                      required
-                      className={`w-full px-4 py-3 pl-10 border rounded-lg shadow-sm focus:outline-none focus:ring-2 focus:ring-blue-500 transition-colors ${
-                        fieldErrors.prenom ? 'border-red-300 focus:border-red-500' : 'border-gray-300 focus:border-blue-500'
-                      }`}
-                      value={formData.prenom}
-                      onChange={handleChange}
-                      onBlur={handleBlur}
-                      placeholder="Votre prénom"
-                    />
-                    <User className="absolute left-3 top-3.5 h-5 w-5 text-gray-400" />
-                  </div>
-                  {fieldErrors.prenom && (
-                    <p className="mt-1 text-sm text-red-600">{fieldErrors.prenom}</p>
-                  )}
-                </div>
-              </div>
-
-              {/* Email */}
-              <div>
-                <label htmlFor="email" className="block text-sm font-medium text-gray-700 mb-2">
-                  Adresse email *
-                </label>
-                <div className="relative">
-                  <input
-                    id="email"
-                    name="email"
-                    type="email"
-                    required
-                    className={`w-full px-4 py-3 pl-10 border rounded-lg shadow-sm focus:outline-none focus:ring-2 focus:ring-blue-500 transition-colors ${
-                      fieldErrors.email ? 'border-red-300 focus:border-red-500' : 'border-gray-300 focus:border-blue-500'
-                    }`}
-                    value={formData.email}
-                    onChange={handleChange}
-                    onBlur={handleBlur}
-                    placeholder="exemple@email.com"
-                  />
-                  <Mail className="absolute left-3 top-3.5 h-5 w-5 text-gray-400" />
-                </div>
-                {fieldErrors.email && (
-                  <p className="mt-1 text-sm text-red-600">{fieldErrors.email}</p>
-                )}
-              </div>
-
-              {/* Téléphone */}
-              <div>
-                <label htmlFor="telephone" className="block text-sm font-medium text-gray-700 mb-2">
-                  Numéro de téléphone *
-                </label>
-                <div className="relative">
-                  <input
-                    id="telephone"
-                    name="telephone"
-                    type="tel"
-                    required
-                    className={`w-full px-4 py-3 pl-10 border rounded-lg shadow-sm focus:outline-none focus:ring-2 focus:ring-blue-500 transition-colors ${
-                      fieldErrors.telephone ? 'border-red-300 focus:border-red-500' : 'border-gray-300 focus:border-blue-500'
-                    }`}
-                    value={formData.telephone}
-                    onChange={handleChange}
-                    onBlur={handleBlur}
-                    placeholder="90000000"
-                  />
-                  <Phone className="absolute left-3 top-3.5 h-5 w-5 text-gray-400" />
-                </div>
-                {fieldErrors.telephone && (
-                  <p className="mt-1 text-sm text-red-600">{fieldErrors.telephone}</p>
-                )}
-              </div>
-
-              {/* Mot de passe */}
-              <div>
-                <label htmlFor="password" className="block text-sm font-medium text-gray-700 mb-2">
-                  Mot de passe *
-                </label>
-                <div className="relative">
-                  <input
-                    id="password"
-                    name="password"
-                    type={showPassword ? 'text' : 'password'}
-                    required
-                    className={`w-full px-4 py-3 pl-10 pr-10 border rounded-lg shadow-sm focus:outline-none focus:ring-2 focus:ring-blue-500 transition-colors ${
-                      fieldErrors.password ? 'border-red-300 focus:border-red-500' : 'border-gray-300 focus:border-blue-500'
-                    }`}
-                    value={formData.password}
-                    onChange={handleChange}
-                    onBlur={handleBlur}
-                    placeholder="••••••••"
-                  />
-                  <Lock className="absolute left-3 top-3.5 h-5 w-5 text-gray-400" />
-                  <button
-                    type="button"
-                    className="absolute right-3 top-3.5 text-gray-400 hover:text-gray-600"
-                    onClick={() => setShowPassword(!showPassword)}
-                  >
-                    {showPassword ? <EyeOff className="h-5 w-5" /> : <Eye className="h-5 w-5" />}
-                  </button>
-                </div>
-                
-                {/* Indicateur de force du mot de passe */}
-                {formData.password && (
-                  <div className="mt-2">
-                    <div className="flex items-center space-x-2">
-                      <div className="flex-1 bg-gray-200 rounded-full h-2">
-                        <div
-                          className={`h-2 rounded-full transition-all duration-300 ${passwordStrength.color}`}
-                          style={{ width: `${(passwordStrength.score / 5) * 100}%` }}
-                        ></div>
-                      </div>
-                      <span className="text-xs text-gray-600">{passwordStrength.label}</span>
-                    </div>
-                  </div>
-                )}
-                
-                {fieldErrors.password && (
-                  <p className="mt-1 text-sm text-red-600">{fieldErrors.password}</p>
-                )}
-              </div>
-
-              {/* Confirmation mot de passe */}
-              <div>
-                <label htmlFor="confirmPassword" className="block text-sm font-medium text-gray-700 mb-2">
-                  Confirmer le mot de passe *
-                </label>
-                <div className="relative">
-                  <input
-                    id="confirmPassword"
-                    name="confirmPassword"
-                    type={showConfirmPassword ? 'text' : 'password'}
-                    required
-                    className={`w-full px-4 py-3 pl-10 pr-10 border rounded-lg shadow-sm focus:outline-none focus:ring-2 focus:ring-blue-500 transition-colors ${
-                      fieldErrors.confirmPassword ? 'border-red-300 focus:border-red-500' : 'border-gray-300 focus:border-blue-500'
-                    }`}
-                    value={formData.confirmPassword}
-                    onChange={handleChange}
-                    onBlur={handleBlur}
-                    placeholder="••••••••"
-                  />
-                  <Lock className="absolute left-3 top-3.5 h-5 w-5 text-gray-400" />
-                  <button
-                    type="button"
-                    className="absolute right-3 top-3.5 text-gray-400 hover:text-gray-600"
-                    onClick={() => setShowConfirmPassword(!showConfirmPassword)}
-                  >
-                    {showConfirmPassword ? <EyeOff className="h-5 w-5" /> : <Eye className="h-5 w-5" />}
-                  </button>
-                </div>
-                {fieldErrors.confirmPassword && (
-                  <p className="mt-1 text-sm text-red-600">{fieldErrors.confirmPassword}</p>
-                )}
-              </div>
-
-              {/* Bouton de soumission */}
-              <button
-                type="submit"
-                disabled={loading}
-                className="w-full bg-gradient-to-r from-blue-600 to-blue-700 text-white py-3 px-4 rounded-lg font-medium hover:from-blue-700 hover:to-blue-800 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 disabled:opacity-50 disabled:cursor-not-allowed transition-all duration-200 transform hover:scale-[1.02]"
-              >
-                {loading ? (
-                  <div className="flex items-center justify-center">
-                    <div className="animate-spin rounded-full h-5 w-5 border-b-2 border-white mr-2"></div>
-                    Création en cours...
-                  </div>
-                ) : (
-                  'Créer mon compte'
-                )}
-              </button>
-
-              {/* Lien de connexion */}
-              <div className="text-center pt-4 border-t border-gray-200">
-                <p className="text-sm text-gray-600">
-                  Déjà un compte ?{' '}
-                  <Link 
-                    href="/login" 
-                    className="font-medium text-blue-600 hover:text-blue-500 transition-colors"
-                  >
-                    Se connecter
-                  </Link>
-                </p>
-              </div>
-            </form>
-          </div>
-        </div>
-      </div>
+      <RegisterForm />
     </div>
   );
 }
