+++ frontend/src/app/types/loginForms.ts
@@ -1,0 +1,17 @@
+
+
+export interface RegisterFormData {
+  nom: string;
+  prenom: string;
+  email: string;
+  password: string;
+  confirmPassword: string;
+  telephone: string;
+  role: "Etudiant" | "Enseignant" | "Responsable";
+}
+
+export type PasswordStrength = {
+  score: number;
+  label: string;
+  color: string;
+};
